CLOSED - # 532: Create utility to import measurements from irouter logs https://verticon-trac.sourcerepo.com/verticon_Tracker/ticket/532

+++ tracker/trunk/com.verticon.tracker.editor/src/com/verticon/tracker/editor/presentation/SampleTrackerProjectWizard.java
@@ -454,7 +454,7 @@ public class SampleTrackerProjectWizard extends BasicNewResourceWizard {
 		return swine;
 	}
 
-	static int getWeigh(int mean){
+	static double getWeigh(int mean){
 		Random r = new Random();
 		return r.nextInt(100) + mean;
 	}
